requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
google-generativeai>=0.8.0,<0.9.0
python-dotenv==1.0.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4